refactor: parser crate and parse calls in ir/codegen
- exposes fewer things beyond the module (makes lexer and lalrpop private)
- moves lexer, parser, ast modules to directories to prepare for adding more functionality cleanly
