+++ parser/src/parser/mod.rs
@@ -1,0 +1,6 @@
+lalrpop_mod!(pub grammar);
+
+pub use grammar::SourceParser;
+
+#[cfg(test)]
+mod tests;
